libgweather/gweather-private.c: Fix compilation in i686; Use guint instead gsize
From https://docs.gtk.org/glib/struct.PtrArray.html, seems the
"len" parameter in the GPtrArray struct is a guint

+++ libgweather/gweather-private.c
@@ -16,14 +16,14 @@ static GWeatherDb *world_db;
 void
 _gweather_location_reset_world (void)
 {
-    gsize i;
+    guint i;
 
     g_return_if_fail (world_db != NULL);
 
     /* At this point, we had a leak if the caches are not completely empty. */
     for (i = 0; i < world_db->locations->len; i++) {
         if (G_UNLIKELY (g_ptr_array_index (world_db->locations, i) != NULL)) {
-            g_warning ("Location with index %li and name %s is still referenced!",
+            g_warning ("Location with index %u and name %s is still referenced!",
                        i,
                        gweather_location_get_name (g_ptr_array_index (world_db->locations, i)));
             g_assert_not_reached ();
@@ -31,7 +31,7 @@ _gweather_location_reset_world (void)
     }
     for (i = 0; i < world_db->timezones->len; i++) {
         if (G_UNLIKELY (g_ptr_array_index (world_db->timezones, i) != NULL)) {
-            g_warning ("Timezone with index %li and tzid %s is still referenced!",
+            g_warning ("Timezone with index %u and tzid %s is still referenced!",
                        i,
                        g_time_zone_get_identifier (g_ptr_array_index (world_db->timezones, i)));
             g_assert_not_reached ();
